Controllers and Services created

+++ backend/src/main/java/com/dio/santander/banklineapi/repositories/TransitionRepository.java
@@ -4,6 +4,4 @@ import com.dio.santander.banklineapi.models.entities.Transition;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransitionRepository extends JpaRepository<Transition, Integer> {
-
-}
+public interface TransitionRepository extends JpaRepository<Transition, Long> {}
